requirements.txt
# FastAPI y dependencias principales
fastapi==0.104.1
uvicorn==0.34.3
python-dotenv==1.1.0
pydantic-settings==2.6.1

# Base de datos
sqlalchemy==2.0.41
greenlet==3.2.3

# Seguridad
passlib==1.7.4
python-jose==3.5.0
pyasn1==0.6.1
rsa==4.9.1
ecdsa==0.19.1
six==1.17.0

# Utilidades
annotated-types==0.7.0
click==8.2.1
h11==0.16.0
colorama==0.4.6

# Dependencias adicionales
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi-cli==0.0.7
httpcore==1.0.8
httptools==0.6.4
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.3.0
orjson==3.10.16
pyasn1-modules==0.4.8
pydantic==2.4.2
pydantic-core==2.33.1
pydantic-extra-types==2.10.3
python-multipart==0.0.6
PyYAML==6.0.2
pyodbc
pytest==7.4.3
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.4.0
watchfiles==1.0.5
websockets==15.0.1
aiofiles==23.2.1
alembic==1.13.1
